Move staging behavior settings to profile settings
The staging behavior configuration UI and logic have been relocated from project appearance settings to profile lanes and branches settings. This centralizes staging behavior preferences under profile settings for improved organization.
