+++ apps/desktop/src/components/profileSettings/LanesAndBranchesSettings.svelte
@@ -1,13 +1,20 @@
 <script lang="ts">
 	import {
 		autoSelectBranchNameFeature,
-		autoSelectBranchCreationFeature
+		autoSelectBranchCreationFeature,
+		stagingBehaviorFeature,
+		type StagingBehavior
 	} from '$lib/config/uiFeatureFlags';
 	import { persisted } from '@gitbutler/shared/persisted';
-	import { CardGroup, Toggle } from '@gitbutler/ui';
+	import { CardGroup, RadioButton, Toggle } from '@gitbutler/ui';
 
 	const addToLeftmost = persisted<boolean>(false, 'branch-placement-leftmost');
-</script>
+	function onStagingBehaviorFormChange(form: HTMLFormElement) {
+		const formData = new FormData(form);
+		const selectedStagingBehavior = formData.get('stagingBehaviorType') as StagingBehavior | null;
+		if (!selectedStagingBehavior) return;
+		stagingBehaviorFeature.set(selectedStagingBehavior);
+	}</script>
 
 <CardGroup.Item standalone labelFor="add-leftmost">
 	{#snippet title()}
@@ -59,4 +66,65 @@
 			/>
 		{/snippet}
 	</CardGroup.Item>
+</CardGroup>
+
+<CardGroup>
+	<form class="stack-v" onchange={(e) => onStagingBehaviorFormChange(e.currentTarget)}>
+		<CardGroup.Item labelFor="stage-all">
+			{#snippet title()}
+				Stage all files
+			{/snippet}
+			{#snippet caption()}
+				Stage all files assigned to the stack on commit. If no files are staged, all unassigned
+				files will be staged.
+			{/snippet}
+			{#snippet actions()}
+				<RadioButton
+					name="stagingBehaviorType"
+					value="all"
+					id="stage-all"
+					checked={$stagingBehaviorFeature === 'all'}
+				/>
+			{/snippet}
+		</CardGroup.Item>
+
+		<CardGroup.Item labelFor="stage-selection">
+			{#snippet title()}
+				Stage selected files
+			{/snippet}
+			{#snippet caption()}
+				Stage the selected assigned files to the stack on commit. If no files are selected, stage
+				all files. If there are no assigned files, stage all selected unassigned files.
+				<br />
+				And if no files are selected, stage all unassigned files.
+			{/snippet}
+			{#snippet actions()}
+				<RadioButton
+					name="stagingBehaviorType"
+					value="selection"
+					id="stage-selection"
+					checked={$stagingBehaviorFeature === 'selection'}
+				/>
+			{/snippet}
+		</CardGroup.Item>
+
+		<CardGroup.Item labelFor="stage-none">
+			{#snippet title()}
+				Don't stage files automatically
+			{/snippet}
+			{#snippet caption()}
+				Do not stage any files automatically.
+				<br />
+				You're more of a DIY developer in that way.
+			{/snippet}
+			{#snippet actions()}
+				<RadioButton
+					name="stagingBehaviorType"
+					value="none"
+					id="stage-none"
+					checked={$stagingBehaviorFeature === 'none'}
+				/>
+			{/snippet}
+		</CardGroup.Item>
+	</form>
 </CardGroup>
