+++ apps/desktop/src/components/projectSettings/AppearanceSettings.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import ThemeSelector from '$components/ThemeSelector.svelte';
-	import { stagingBehaviorFeature, type StagingBehavior } from '$lib/config/uiFeatureFlags';
 	import { SETTINGS } from '$lib/settings/userSettings';
 	import { inject } from '@gitbutler/core/context';
 	import {
@@ -38,13 +37,6 @@
 			...s,
 			scrollbarVisibilityState: selectedScrollbarVisibility
 		}));
-	}
-
-	function onStagingBehaviorFormChange(form: HTMLFormElement) {
-		const formData = new FormData(form);
-		const selectedStagingBehavior = formData.get('stagingBehaviorType') as StagingBehavior | null;
-		if (!selectedStagingBehavior) return;
-		stagingBehaviorFeature.set(selectedStagingBehavior);
 	}
 </script>
 
@@ -339,67 +331,6 @@
 					value="always"
 					id="scrollbar-always"
 					checked={$userSettings.scrollbarVisibilityState === 'always'}
-				/>
-			{/snippet}
-		</CardGroup.Item>
-	</form>
-</CardGroup>
-
-<CardGroup>
-	<form class="stack-v" onchange={(e) => onStagingBehaviorFormChange(e.currentTarget)}>
-		<CardGroup.Item labelFor="stage-all">
-			{#snippet title()}
-				Stage all files
-			{/snippet}
-			{#snippet caption()}
-				Stage all files assigned to the stack on commit. If no files are staged, all unassigned
-				files will be staged.
-			{/snippet}
-			{#snippet actions()}
-				<RadioButton
-					name="stagingBehaviorType"
-					value="all"
-					id="stage-all"
-					checked={$stagingBehaviorFeature === 'all'}
-				/>
-			{/snippet}
-		</CardGroup.Item>
-
-		<CardGroup.Item labelFor="stage-selection">
-			{#snippet title()}
-				Stage selected files
-			{/snippet}
-			{#snippet caption()}
-				Stage the selected assigned files to the stack on commit. If no files are selected, stage
-				all files. If there are no assigned files, stage all selected unassigned files.
-				<br />
-				And if no files are selected, stage all unassigned files.
-			{/snippet}
-			{#snippet actions()}
-				<RadioButton
-					name="stagingBehaviorType"
-					value="selection"
-					id="stage-selection"
-					checked={$stagingBehaviorFeature === 'selection'}
-				/>
-			{/snippet}
-		</CardGroup.Item>
-
-		<CardGroup.Item labelFor="stage-none">
-			{#snippet title()}
-				Don't stage files automatically
-			{/snippet}
-			{#snippet caption()}
-				Do not stage any files automatically.
-				<br />
-				You're more of a DIY developer in that way.
-			{/snippet}
-			{#snippet actions()}
-				<RadioButton
-					name="stagingBehaviorType"
-					value="none"
-					id="stage-none"
-					checked={$stagingBehaviorFeature === 'none'}
 				/>
 			{/snippet}
 		</CardGroup.Item>
